Added basics of VM and stack and basic arithmetic operation

+++ chunk.c
@@ -22,7 +22,7 @@ void write_chunk(Chunk* chunk, uint8_t byte, int line) {
         int old_capacity = chunk->capacity;
         chunk->capacity = GROW_CAPACITY(old_capacity);
         chunk->code = GROW_ARRAY(uint8_t, chunk->code, old_capacity, chunk->capacity);
-        chunk->lines = GROW_ARRAY(uint8_t, chunk->lines, old_capacity, chunk->capacity);
+        chunk->lines = GROW_ARRAY(int, chunk->lines, old_capacity, chunk->capacity);
     }
 
     chunk->code[chunk->count] = byte;
@@ -40,6 +40,7 @@ void free_chunk(Chunk* chunk) {
 }
 
 int add_constant(Chunk* chunk, Value value) {
+    // Add the value to the constant pool
     write_value_array(&chunk->constants, value);
     // We then return the index of the value
     return chunk->constants.count - 1;
